# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
...
  /api/login:
    post:
      summary: 'User Login'
      operationId: userLogin
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  user:
                    id: 1
                    name: 'John Doe'
                    email: user@example.com
                    email_verified_at: '2021-03-01T00:00:00.000000Z'
                    created_at: '2021-03-01T00:00:00.000000Z'
                    updated_at: '2021-03-01T00:00:00.000000Z'
                  access_token: eyJ0eXAiOiJKV1QiLCJh...
                  token_type: Bearer
                  expires_at: '2021-03-01T00:00:00.000000Z'
                properties:
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      name:
                        type: string
                        example: 'John Doe'
                      email:
                        type: string
                        example: user@example.com
                      email_verified_at:
                        type: string
                        example: '2021-03-01T00:00:00.000000Z'
                      created_at:
                        type: string
                        example: '2021-03-01T00:00:00.000000Z'
                      updated_at:
                        type: string
                        example: '2021-03-01T00:00:00.000000Z'
                  access_token:
                    type: string
                    example: eyJ0eXAiOiJKV1QiLCJh...
                  token_type:
                    type: string
                    example: Bearer
                  expires_at:
                    type: string
                    example: '2021-03-01T00:00:00.000000Z'
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: 'The email of the user.'
                  example: user@example.com
                password:
                  type: string
                  description: 'The password of the user.'
                  example: secret
              required:
                - email
                - password
      security: []
  /api/reset-password:
    post:
      summary: 'Reset Password'
      operationId: resetPassword
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Password reset successfully'
                properties:
                  message:
                    type: string
                    example: 'Password reset successfully'
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: 'The email of the user.'
                  example: user@example.com
                password:
                  type: string
                  description: 'The new password of the user.'
                  example: newpassword
              required:
                - email
                - password
      security: []
  /api/refresh-token:
    post:
      summary: 'Refresh Token'
      operationId: refreshToken
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  access_token: new_access_token
                  access_token_expires_at: '2021-03-01T00:00:00.000000Z'
                properties:
                  access_token:
                    type: string
                    example: new_access_token
                  access_token_expires_at:
                    type: string
                    example: '2021-03-01T00:00:00.000000Z'
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
                  description: 'The refresh token of the user.'
                  example: old_refresh_token
              required:
                - refresh_token
      security: []
...